requirement.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
cycler==0.10.0
decorator==4.3.0
future==0.17.1
idna==2.7
imageio==2.4.1
Khayyam3==1.0.0
kiwisolver==1.0.1
matplotlib==3.0.1
moviepy==0.2.3.5
numpy==1.15.4
Pillow==5.3.0
psutil==5.4.8
pycparser==2.19
pyparsing==2.3.0
python-dateutil==2.7.5
python-telegram-bot==11.1.0
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.13
tqdm==4.28.1
urllib3==1.24.1
